requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dotenv
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rsconnect_python==1.26.0
safetensors==0.5.3
semver==3.0.4
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
timm==1.0.15
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3